task 0 fix error

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int n = 0;
+	size_t n = 0;
 
 	if (h == NULL)
 	return (-1);
